+++ packages/interface-compliance-tests/src/transport/dial-test.ts
@@ -1,4 +1,4 @@
-import { AbortError, TypedEventEmitter } from '@libp2p/interface'
+import { TypedEventEmitter } from '@libp2p/interface'
 import { expect } from 'aegir/chai'
 import all from 'it-all'
 import drain from 'it-drain'
@@ -19,9 +19,10 @@ export default (common: TestSetup<TransportTestFixtures>): void => {
     let registrar: Registrar
     let listenAddrs: Multiaddr[]
     let dialAddrs: Multiaddr[]
-    let transport: Transport
+    let dialer: Transport
+    let listener: Transport
     let connector: Connector
-    let listener: Listener
+    let listen: Listener
 
     before(async () => {
       registrar = mockRegistrar()
@@ -30,7 +31,7 @@ export default (common: TestSetup<TransportTestFixtures>): void => {
         events: new TypedEventEmitter()
       });
 
-      ({ listenAddrs, dialAddrs, transport, connector } = await common.setup())
+      ({ listenAddrs, dialAddrs, dialer, listener, connector } = await common.setup())
     })
 
     after(async () => {
@@ -38,16 +39,16 @@ export default (common: TestSetup<TransportTestFixtures>): void => {
     })
 
     beforeEach(async () => {
-      listener = transport.createListener({
+      listen = listener.createListener({
         upgrader
       })
-      await listener.listen(listenAddrs[0])
+      await listen.listen(listenAddrs[0])
     })
 
     afterEach(async () => {
       sinon.restore()
       connector.restore()
-      await listener.close()
+      await listen.close()
     })
 
     it('simple', async () => {
@@ -62,7 +63,7 @@ export default (common: TestSetup<TransportTestFixtures>): void => {
       })
 
       const upgradeSpy = sinon.spy(upgrader, 'upgradeOutbound')
-      const conn = await transport.dial(dialAddrs[0], {
+      const conn = await dialer.dial(dialAddrs[0], {
         upgrader
       })
 
@@ -78,7 +79,7 @@ export default (common: TestSetup<TransportTestFixtures>): void => {
 
     it('can close connections', async () => {
       const upgradeSpy = sinon.spy(upgrader, 'upgradeOutbound')
-      const conn = await transport.dial(dialAddrs[0], {
+      const conn = await dialer.dial(dialAddrs[0], {
         upgrader
       })
 
@@ -91,7 +92,7 @@ export default (common: TestSetup<TransportTestFixtures>): void => {
     it('to non existent listener', async () => {
       const upgradeSpy = sinon.spy(upgrader, 'upgradeOutbound')
 
-      await expect(transport.dial(dialAddrs[1], {
+      await expect(dialer.dial(dialAddrs[1], {
         upgrader
       })).to.eventually.be.rejected()
       expect(upgradeSpy.callCount).to.equal(0)
@@ -101,9 +102,9 @@ export default (common: TestSetup<TransportTestFixtures>): void => {
       const upgradeSpy = sinon.spy(upgrader, 'upgradeOutbound')
       const controller = new AbortController()
       controller.abort()
-      const conn = transport.dial(dialAddrs[0], { signal: controller.signal, upgrader })
+      const conn = dialer.dial(dialAddrs[0], { signal: controller.signal, upgrader })
 
-      await expect(conn).to.eventually.be.rejected().with.property('code', AbortError.code)
+      await expect(conn).to.eventually.be.rejected().with.property('name', 'AbortError')
       expect(upgradeSpy.callCount).to.equal(0)
     })
 
@@ -114,10 +115,10 @@ export default (common: TestSetup<TransportTestFixtures>): void => {
       connector.delay(100)
 
       const controller = new AbortController()
-      const conn = transport.dial(dialAddrs[0], { signal: controller.signal, upgrader })
+      const conn = dialer.dial(dialAddrs[0], { signal: controller.signal, upgrader })
       setTimeout(() => { controller.abort() }, 50)
 
-      await expect(conn).to.eventually.be.rejected().with.property('code', AbortError.code)
+      await expect(conn).to.eventually.be.rejected().with.property('name', 'AbortError')
       expect(upgradeSpy.callCount).to.equal(0)
     })
   })