+++ packages/transport-webrtc/src/private-to-public/transport.ts
@@ -1,5 +1,5 @@
 import { noise } from '@chainsafe/libp2p-noise'
-import { type CreateListenerOptions, transportSymbol, type Transport, type Listener, type ComponentLogger, type Logger, type Connection, type CounterGroup, type Metrics, type PeerId } from '@libp2p/interface'
+import { transportSymbol, serviceCapabilities } from '@libp2p/interface'
 import * as p from '@libp2p/peer-id'
 import { protocols } from '@multiformats/multiaddr'
 import { WebRTCDirect } from '@multiformats/multiaddr-matcher'
@@ -17,6 +17,7 @@ import * as sdp from './sdp.js'
 import { genUfrag } from './util.js'
 import type { WebRTCDialOptions } from './options.js'
 import type { DataChannelOptions } from '../index.js'
+import type { CreateListenerOptions, Transport, Listener, ComponentLogger, Logger, Connection, CounterGroup, Metrics, PeerId } from '@libp2p/interface'
 import type { Multiaddr } from '@multiformats/multiaddr'
 
 /**
@@ -52,6 +53,7 @@ export interface WebRTCMetrics {
 }
 
 export interface WebRTCTransportDirectInit {
+  rtcConfiguration?: RTCConfiguration | (() => RTCConfiguration | Promise<RTCConfiguration>)
   dataChannel?: DataChannelOptions
 }
 
@@ -73,6 +75,14 @@ export class WebRTCDirectTransport implements Transport {
       }
     }
   }
+
+  readonly [transportSymbol] = true
+
+  readonly [Symbol.toStringTag] = '@libp2p/webrtc-direct'
+
+  readonly [serviceCapabilities]: string[] = [
+    '@libp2p/transport'
+  ]
 
   /**
    * Dial a given multiaddr
@@ -105,16 +115,6 @@ export class WebRTCDirectTransport implements Transport {
   }
 
   /**
-   * Implement toString() for WebRTCTransport
-   */
-  readonly [Symbol.toStringTag] = '@libp2p/webrtc-direct'
-
-  /**
-   * Symbol.for('@libp2p/transport')
-   */
-  readonly [transportSymbol] = true
-
-  /**
    * Connect to a peer using a multiaddr
    */
   async _connect (ma: Multiaddr, options: WebRTCDialOptions): Promise<Connection> {
@@ -139,7 +139,10 @@ export class WebRTCDirectTransport implements Transport {
       hash: sdp.toSupportedHashFunction(remoteCerthash.code)
     } as any)
 
-    const peerConnection = new RTCPeerConnection({ certificates: [certificate] })
+    const peerConnection = new RTCPeerConnection({
+      ...(typeof this.init.rtcConfiguration === 'function' ? await this.init.rtcConfiguration() : this.init.rtcConfiguration ?? {}),
+      certificates: [certificate]
+    })
 
     try {
       // create data channel for running the noise handshake. Once the data channel is opened,