+++ packages/interface-internal/CHANGELOG.md
@@ -32,6 +32,17 @@
   * dependencies
     * @libp2p/peer-collections bumped from ^5.1.3 to ^5.1.4
 
+## [1.2.3](https://github.com/libp2p/js-libp2p/compare/interface-internal-v1.2.2...interface-internal-v1.2.3) (2024-06-07)
+
+
+### Dependencies
+
+* bump aegir from 42.2.11 to 43.0.1 ([#2571](https://github.com/libp2p/js-libp2p/issues/2571)) ([757fb26](https://github.com/libp2p/js-libp2p/commit/757fb2674f0a3e06fd46d3ff63f7f461c32d47d2))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.4.0 to ^1.4.1
+    * @libp2p/peer-collections bumped from ^5.2.2 to ^5.2.3
+
 ## [1.2.2](https://github.com/libp2p/js-libp2p/compare/interface-internal-v1.2.1...interface-internal-v1.2.2) (2024-05-17)