+++ packages/metrics-devtools/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.1.1](https://github.com/libp2p/js-libp2p/compare/devtools-metrics-v0.1.0...devtools-metrics-v0.1.1) (2024-06-07)
+
+
+### Dependencies
+
+* bump aegir from 42.2.11 to 43.0.1 ([#2571](https://github.com/libp2p/js-libp2p/issues/2571)) ([757fb26](https://github.com/libp2p/js-libp2p/commit/757fb2674f0a3e06fd46d3ff63f7f461c32d47d2))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.3.1 to ^1.4.1
+    * @libp2p/interface-internal bumped from ^1.2.1 to ^1.2.3
+    * @libp2p/logger bumped from ^4.0.12 to ^4.0.14
+    * @libp2p/simple-metrics bumped from ^1.0.1 to ^1.0.3
+  * devDependencies
+    * @libp2p/peer-id-factory bumped from ^4.1.1 to ^4.1.3
+
 ## [0.1.0](https://github.com/libp2p/js-libp2p/compare/devtools-metrics-v0.0.1...devtools-metrics-v0.1.0) (2024-05-28)