+++ packages/metrics-simple/CHANGELOG.md
@@ -5,6 +5,17 @@
 
 * track stream metrics ([#2](https://github.com/libp2p/js-libp2p-simple-metrics/issues/2)) ([caafb3d](https://github.com/libp2p/js-libp2p-simple-metrics/commit/caafb3d103fd7df0a2a4e6b3e800f4bc9c35c58f))
 
+## [1.0.3](https://github.com/libp2p/js-libp2p/compare/simple-metrics-v1.0.2...simple-metrics-v1.0.3) (2024-06-07)
+
+
+### Dependencies
+
+* bump aegir from 42.2.11 to 43.0.1 ([#2571](https://github.com/libp2p/js-libp2p/issues/2571)) ([757fb26](https://github.com/libp2p/js-libp2p/commit/757fb2674f0a3e06fd46d3ff63f7f461c32d47d2))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.4.0 to ^1.4.1
+    * @libp2p/logger bumped from ^4.0.13 to ^4.0.14
+
 ## [1.0.2](https://github.com/libp2p/js-libp2p/compare/simple-metrics-v1.0.1...simple-metrics-v1.0.2) (2024-05-17)