+++ packages/multistream-select/CHANGELOG.md
@@ -19,6 +19,18 @@
   * devDependencies
     * @libp2p/logger bumped from ^4.0.1 to ^4.0.2
 
+## [5.1.11](https://github.com/libp2p/js-libp2p/compare/multistream-select-v5.1.10...multistream-select-v5.1.11) (2024-06-07)
+
+
+### Dependencies
+
+* bump aegir from 42.2.11 to 43.0.1 ([#2571](https://github.com/libp2p/js-libp2p/issues/2571)) ([757fb26](https://github.com/libp2p/js-libp2p/commit/757fb2674f0a3e06fd46d3ff63f7f461c32d47d2))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.4.0 to ^1.4.1
+  * devDependencies
+    * @libp2p/logger bumped from ^4.0.13 to ^4.0.14
+
 ## [5.1.10](https://github.com/libp2p/js-libp2p/compare/multistream-select-v5.1.9...multistream-select-v5.1.10) (2024-05-17)