+++ packages/protocol-echo/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [1.0.8](https://github.com/libp2p/js-libp2p/compare/echo-v1.0.7...echo-v1.0.8) (2024-06-07)
+
+
+### Dependencies
+
+* bump aegir from 42.2.11 to 43.0.1 ([#2571](https://github.com/libp2p/js-libp2p/issues/2571)) ([757fb26](https://github.com/libp2p/js-libp2p/commit/757fb2674f0a3e06fd46d3ff63f7f461c32d47d2))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^1.4.0 to ^1.4.1
+    * @libp2p/interface-internal bumped from ^1.2.2 to ^1.2.3
+  * devDependencies
+    * @libp2p/logger bumped from ^4.0.13 to ^4.0.14
+
 ## [1.0.7](https://github.com/libp2p/js-libp2p/compare/echo-v1.0.6...echo-v1.0.7) (2024-05-17)